Added in all the sounds to the code.
I changed all the names and stuff for your part so it should work if you
comment out my file paths.

We are missing notes for the flats of all the notes. We are good on the
sharps and regular notes.

Added a stopPlaying function at the end that we can invoke. I want to be
able to use it once the user takes their finger off a button.